// src/common/enums/warehouse.enum.ts
export enum warehouseStatusType {
  PENDING = "pending",
  CREATED = "created",
  FAILED = "failed",
  INACTIVE = "inactive",
  REQUEST_INITIATED = "request initiated",
  REMOVAL_PENDING = "removal pending",
}

export enum warehouseJobTypeEnums {
  REGISTER_WAREHOUSE = "register warehouse",
  REMOVE_WAREHOUSE = "remove warehouse",
}